+++ node/src/components/consensus.rs
@@ -9,11 +9,13 @@ mod era_supervisor;
 #[macro_use]
 mod highway_core;
 pub(crate) mod error;
+mod leader_sequence;
 mod metrics;
 mod protocols;
 #[cfg(test)]
 pub(crate) mod tests;
 mod traits;
+pub(crate) mod utils;
 mod validator_change;
 
 use std::{
@@ -25,7 +27,6 @@ use std::{
 
 use datasize::DataSize;
 use derive_more::From;
-use hex_fmt::HexFmt;
 use serde::{Deserialize, Serialize};
 use tracing::{info, trace};
 
@@ -36,7 +37,7 @@ use crate::{
     effect::{
         announcements::{ConsensusAnnouncement, FatalAnnouncement, PeerBehaviorAnnouncement},
         diagnostics_port::DumpConsensusStateRequest,
-        incoming::ConsensusMessageIncoming,
+        incoming::{ConsensusDemand, ConsensusMessageIncoming},
         requests::{
             BlockValidationRequest, ChainspecRawBytesRequest, ConsensusRequest,
             ContractRuntimeRequest, DeployBufferRequest, NetworkInfoRequest, NetworkRequest,
@@ -49,6 +50,8 @@ use crate::{
     types::{BlockHash, BlockHeader, BlockPayload, NodeId},
     NodeRng,
 };
+use protocols::{highway::HighwayProtocol, zug::Zug};
+use traits::Context;
 
 pub(crate) use cl_context::ClContext;
 pub(crate) use config::{ChainspecConsensusExt, Config};
@@ -56,18 +59,86 @@ pub(crate) use consensus_protocol::{BlockContext, EraReport, ProposedBlock};
 pub(crate) use era_supervisor::{debug::EraDump, EraSupervisor};
 #[cfg(test)]
 pub(crate) use highway_core::highway::Vertex as HighwayVertex;
-#[cfg(test)]
+pub(crate) use leader_sequence::LeaderSequence;
 pub(crate) use protocols::highway::HighwayMessage;
-pub(crate) use protocols::highway::HighwayProtocol;
 pub(crate) use validator_change::ValidatorChange;
+
+/// A message to be handled by the consensus protocol instance in a particular era.
+#[derive(DataSize, Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
+pub(crate) enum EraMessage<C>
+where
+    C: Context,
+{
+    Zug(Box<protocols::zug::Message<C>>),
+    Highway(Box<HighwayMessage<C>>),
+}
+
+impl<C: Context> EraMessage<C> {
+    /// Returns the message for the Zug protocol, or an error if it is for a different protocol.
+    fn try_into_zug(self) -> Result<protocols::zug::Message<C>, Self> {
+        match self {
+            EraMessage::Zug(msg) => Ok(*msg),
+            other => Err(other),
+        }
+    }
+
+    /// Returns the message for the Highway protocol, or an error if it is for a different
+    /// protocol.
+    pub(crate) fn try_into_highway(self) -> Result<HighwayMessage<C>, Self> {
+        match self {
+            EraMessage::Highway(msg) => Ok(*msg),
+            other => Err(other),
+        }
+    }
+}
+
+impl<C: Context> From<protocols::zug::Message<C>> for EraMessage<C> {
+    fn from(msg: protocols::zug::Message<C>) -> EraMessage<C> {
+        EraMessage::Zug(Box::new(msg))
+    }
+}
+
+impl<C: Context> From<HighwayMessage<C>> for EraMessage<C> {
+    fn from(msg: HighwayMessage<C>) -> EraMessage<C> {
+        EraMessage::Highway(Box::new(msg))
+    }
+}
+
+/// A request to be handled by the consensus protocol instance in a particular era.
+#[derive(DataSize, Debug, Clone, Serialize, Deserialize, PartialEq, Eq, From)]
+pub(crate) enum EraRequest<C>
+where
+    C: Context,
+{
+    Zug(protocols::zug::SyncRequest<C>),
+}
+
+impl<C: Context> EraRequest<C> {
+    /// Returns the request for the Zug protocol, or an error if it is for a different protocol.
+    fn try_into_zug(self) -> Result<protocols::zug::SyncRequest<C>, Self> {
+        match self {
+            EraRequest::Zug(msg) => Ok(msg),
+        }
+    }
+}
 
 #[derive(DataSize, Clone, Serialize, Deserialize)]
 pub(crate) enum ConsensusMessage {
     /// A protocol message, to be handled by the instance in the specified era.
-    Protocol { era_id: EraId, payload: Vec<u8> },
+    Protocol {
+        era_id: EraId,
+        payload: EraMessage<ClContext>,
+    },
     /// A request for evidence against the specified validator, from any era that is still bonded
     /// in `era_id`.
     EvidenceRequest { era_id: EraId, pub_key: PublicKey },
+}
+
+/// A protocol request message, to be handled by the instance in the specified era.
+#[derive(DataSize, Clone, Serialize, Deserialize)]
+pub(crate) struct ConsensusRequestMessage {
+    era_id: EraId,
+    payload: EraRequest<ClContext>,
 }
 
 /// An ID to distinguish different timers. What they are used for is specific to each consensus
@@ -98,9 +169,12 @@ pub struct ResolveValidity {
 /// Consensus component event.
 #[derive(DataSize, Debug, From)]
 pub(crate) enum Event {
-    #[from]
     /// An incoming network message.
+    #[from]
     Incoming(ConsensusMessageIncoming),
+    /// An incoming demand message.
+    #[from]
+    DemandIncoming(ConsensusDemand),
     /// A scheduled event to be handled by a specified era.
     Timer {
         era_id: EraId,
@@ -150,7 +224,7 @@ impl Display for ConsensusMessage {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
             ConsensusMessage::Protocol { era_id, payload } => {
-                write!(f, "protocol message {:10} in {}", HexFmt(payload), era_id)
+                write!(f, "protocol message {:?} in {}", payload, era_id)
             }
             ConsensusMessage::EvidenceRequest { era_id, pub_key } => write!(
                 f,
@@ -161,11 +235,30 @@ impl Display for ConsensusMessage {
     }
 }
 
+impl Debug for ConsensusRequestMessage {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "ConsensusRequestMessage {{ era_id: {:?}, .. }}",
+            self.era_id
+        )
+    }
+}
+
+impl Display for ConsensusRequestMessage {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        write!(f, "protocol request {:?} in {}", self.payload, self.era_id)
+    }
+}
+
 impl Display for Event {
     fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
         match self {
             Event::Incoming(ConsensusMessageIncoming { sender, message }) => {
-                write!(f, "msg from {:?}: {}", sender, message)
+                write!(f, "message from {:?}: {}", sender, message)
+            }
+            Event::DemandIncoming(demand) => {
+                write!(f, "demand from {:?}: {}", demand.sender, demand.request_msg)
             }
             Event::Timer {
                 era_id,
@@ -233,6 +326,7 @@ pub(crate) trait ReactorEventT:
     + From<Event>
     + Send
     + From<NetworkRequest<Message>>
+    + From<ConsensusDemand>
     + From<NetworkInfoRequest>
     + From<DeployBufferRequest>
     + From<ConsensusAnnouncement>
@@ -249,6 +343,7 @@ impl<REv> ReactorEventT for REv where
     REv: ReactorEvent
         + From<Event>
         + Send
+        + From<ConsensusDemand>
         + From<NetworkRequest<Message>>
         + From<NetworkInfoRequest>
         + From<DeployBufferRequest>
@@ -287,13 +382,18 @@ where
             Event::Incoming(ConsensusMessageIncoming { sender, message }) => {
                 self.handle_message(effect_builder, rng, sender, message)
             }
+            Event::DemandIncoming(ConsensusDemand {
+                sender,
+                request_msg: demand,
+                auto_closing_responder,
+            }) => self.handle_demand(effect_builder, rng, sender, demand, auto_closing_responder),
             Event::NewBlockPayload(new_block_payload) => {
                 self.handle_new_block_payload(effect_builder, rng, new_block_payload)
             }
             Event::BlockAdded {
                 header,
                 header_hash: _,
-            } => self.handle_block_added(effect_builder, *header),
+            } => self.handle_block_added(effect_builder, rng, *header),
             Event::ResolveValidity(resolve_validity) => {
                 self.resolve_validity(effect_builder, rng, resolve_validity)
             }