+++ node/src/components/consensus/tests/utils.rs
@@ -24,15 +24,19 @@ pub static ALICE_NODE_ID: Lazy<NodeId> = Lazy::new(|| {
     })))
 });
 
-pub static BOB_PRIVATE_KEY: Lazy<SecretKey> =
-    Lazy::new(|| SecretKey::ed25519_from_bytes([1; SecretKey::ED25519_LENGTH]).unwrap());
-pub static BOB_PUBLIC_KEY: Lazy<PublicKey> = Lazy::new(|| PublicKey::from(&*BOB_PRIVATE_KEY));
+pub static BOB_SECRET_KEY: Lazy<Arc<SecretKey>> =
+    Lazy::new(|| Arc::new(SecretKey::ed25519_from_bytes([1; SecretKey::ED25519_LENGTH]).unwrap()));
+pub static BOB_PUBLIC_KEY: Lazy<PublicKey> = Lazy::new(|| PublicKey::from(&**BOB_SECRET_KEY));
 pub static BOB_NODE_ID: Lazy<NodeId> = Lazy::new(|| {
     NodeId::from(KeyFingerprint::from(Sha512::new(match *BOB_PUBLIC_KEY {
         PublicKey::Ed25519(pub_key) => pub_key,
         _ => panic!("BOB_PUBLIC_KEY is Ed25519"),
     })))
 });
+
+pub static CAROL_SECRET_KEY: Lazy<Arc<SecretKey>> =
+    Lazy::new(|| Arc::new(SecretKey::ed25519_from_bytes([2; SecretKey::ED25519_LENGTH]).unwrap()));
+pub static CAROL_PUBLIC_KEY: Lazy<PublicKey> = Lazy::new(|| PublicKey::from(&**CAROL_SECRET_KEY));
 
 /// Loads the local chainspec and overrides timestamp and genesis account with the given stakes.
 /// The test `Chainspec` returned has eras with exactly two blocks.