+++ node/CHANGELOG.md
@@ -53,6 +53,11 @@ All notable changes to this project will be documented in this file.  The format
 * Add `testing` feature to casper-node crate to support test-only functionality (random constructors) on blocks and deploys.
 * The network handshake now contains the hash of the chainspec used and will be successful only if they match.
 * Add an `identity` option to load existing network identity certificates signed by a CA.
+* Make consensus settings non-optional. A value 0 disables them.
+* Add a `[consensus.zug]` section to `config.toml` for the Zug protocol.
+* Add a `consensus_protocol` setting to the chainspec to choose a consensus protocol, and a `minimum_block_time` setting for the minimum difference between a block's timestamp and its child's.
+* Move `finality_threshold_fraction` from the `highway` to the `core` section in the chainspec.
+* Move `max_execution_delay` from the `highway` to the `consensus` section in the `config.toml`.
 
 ### Changed
 * Detection of a crash no longer triggers DB integrity checks to run on node start; the checks can be triggered manually instead.
@@ -311,6 +316,7 @@ All notable changes to this project will be documented in this file.  The format
 * Initialize consensus round success meter with current timestamp.
 * Era Supervisor now accounts for the last emergency restart.
 * Upgrade dependencies, in particular tokio.
+* Use `minimum_block_time` and `maximum_round_length` in Highway, instead of `minimum_round_exponent` and `maximum_round_exponent`. The minimum round length doesn't have to be a power of two in milliseconds anymore.
 
 ### Removed
 * Remove `impl Sub<Timestamp> for Timestamp` to help avoid panicking in non-obvious edge cases.